Add some terraform to spin up infra and start up qortal core
The `qortal_bootstrap.sh` is what does the needful to get core up and
running. It creates a qortal user and runs things under
`/opt/qortal`. It also uses the 2/2021 bootstrap from
crowetic. Eventully I'll see if I can back up the database on a
cadence and store it in an object store and use that bootstrap
instead.