Update the script to use 1.4.2 of core and the qortaldb bootstrap
I'm hosting this myself on spaces as I suspect the download will be
faster than using mediafire. I'll plan on scripting it when updates
are made to the db snapshots.

In trying this out, I did see that it does reset the blockchain to a
lower block. I think that is a good thing generally as it makes sure
we're not forked and we should also get our core versions updated
automatically (I *think*).